requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
dataclasses-json==0.6.7
daytona_sdk
distro==1.9.0
e2b==1.0.5
e2b-code-interpreter==1.0.3
environs==9.5.0
fastapi==0.115.6
frozenlist==1.5.0
grpclib==0.4.7
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.0.1
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langsmith==0.2.4
markdown-it-py==3.0.0
marshmallow==3.23.2
mdurl==0.1.2
modal==0.72.15
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.1
openai==1.58.1
orjson==3.10.12
packaging==24.2
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
scipy==1.15.1
shellingham==1.5.4
sigtools==4.0.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
synchronicity==0.9.8
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
tiktoken==0.8.0
toml==0.10.2
tqdm==4.67.1
typer==0.15.1
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
watchfiles==1.0.4
yarl==1.18.3
